requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Flask-Sijax==0.3.3
Flask-SocketIO==0.5.0
Flask-WTF==0.8.4
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.8
Sijax==0.2.7
Tempita==0.5.2
WTForms==2.0.1
Werkzeug==0.9.6
decorator==3.4.0
gunicorn==19.1.1
itsdangerous==0.24
pbr==0.10.7
six==1.8.0
sqlalchemy-migrate==0.9.2
sqlparse==0.1.14
